ml/room-segmentation/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "openlintel-room-segmentation"
version = "0.1.0"
description = "Room analysis pipeline: VLM detection, SAM 2 segmentation, and depth estimation"
requires-python = ">=3.11"
license = "AGPL-3.0-only"
authors = [{ name = "OpenLintel Team" }]
dependencies = [
    "openlintel-shared",
    "torch>=2.2,<3",
    "torchvision>=0.17,<1",
    "opencv-python-headless>=4.9,<5",
    "numpy>=1.26,<3",
    "pillow>=11.0.0",
    "pydantic>=2.7,<3",
    "litellm>=1.40,<2",
    "structlog>=24.1,<25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
    "pytest-cov>=5,<6",
    "mypy>=1.10,<2",
    "ruff>=0.4,<1",
]
models = [
    "segment-anything-2>=0.1",
    "depth-anything-v2>=0.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/openlintel_segmentation"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.envs.default]
dependencies = [
    "openlintel-shared @ {root:uri}/../../packages/python-shared",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "TCH", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["openlintel_segmentation", "openlintel_shared"]
